requirements.txt
fastapi
uvicorn[standard]
gunicorn
tensorflow
pillow
numpy
python-multipart

# MongoDB dependencies
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.0.0
pydantic[email]
python-dotenv>=1.0.0

# Production dependencies
httpx
aiofiles
